Move verification and deletion modals
