+++ app/src/modals/DeleteCompetitionModal/DeleteCompetitionModal.scss
@@ -1,6 +1,6 @@
-@import '../../../../styles.scss';
+@import '../../styles.scss';
 
-.delete-group {
+.delete-competition {
   position: fixed;
   top: 0;
   left: 0;
