+++ app/src/modals/DeleteCompetitionModal/index.js
