+++ app/src/modals/DeleteGroupModal/DeleteGroupModal.scss
@@ -1,6 +1,6 @@
-@import '../../../../styles.scss';
+@import '../../styles.scss';
 
-.delete-competition {
+.delete-group {
   position: fixed;
   top: 0;
   left: 0;
