+++ app/src/modals/DeleteGroupModal/index.js
