+++ app/src/modals/VerificationModal/VerificationModal.jsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Button from '../../../../components/Button';
+import Button from '../../components/Button';
 import './VerificationModal.scss';
 
-function VerificationModal({ verificationCode, onConfirm }) {
+function VerificationModal({ verificationCode, entity, onConfirm }) {
   return (
     <div className="verification-modal">
       <div className="verification-modal__modal">
-        <span className="code-label">Your group&apos;s verification code is:</span>
+        <span className="code-label">{`Your ${entity}'s verification code is:`}</span>
         <h1 className="code-value">{verificationCode}</h1>
         <p className="code-description">
-          Please save this code somewhere, without it you won&apos;t be able to edit or delete this group
-          in the future.
+          {` Please save this code somewhere, without it you won't \
+             be able to edit or delete this ${entity} in the future.`}
         </p>
         <Button text="Ok, I got it" onClick={onConfirm} />
       </div>
@@ -21,6 +21,7 @@ function VerificationModal({ verificationCode, onConfirm }) {
 
 VerificationModal.propTypes = {
   verificationCode: PropTypes.string.isRequired,
+  entity: PropTypes.string.isRequired,
   onConfirm: PropTypes.func.isRequired
 };
