+++ app/src/modals/VerificationModal/VerificationModal.scss
@@ -1,6 +1,6 @@
-@import '../../../../styles.scss';
+@import '../../styles.scss';
 
-.verification-popup {
+.verification-modal {
   background: rgba(0, 0, 0, 0.5);
   position: fixed;
   top: 0;
