+++ app/src/modals/VerificationModal/index.js
