--- app/src/pages/CreateGroup/components/VerificationModal/VerificationModal.scss
@@ -1,68 +1,0 @@
-@import '../../../../styles.scss';
-
-.verification-modal {
-  background: rgba(0, 0, 0, 0.5);
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  z-index: 500;
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-
-  &__modal {
-    @extend %panel;
-    background: $gray-00;
-    width: 300px;
-    max-width: 90%;
-    min-height: 100px;
-    box-shadow: 0px 5px 10px rgba(0, 0, 0, 0.5);
-    padding: 25px;
-    display: flex;
-    flex-direction: column;
-
-    .code-label {
-      display: block;
-      width: 100%;
-      font-size: 0.8em;
-      text-align: center;
-      color: $brand-color;
-    }
-
-    .code-value {
-      margin: 0;
-      margin-top: 20px;
-      text-align: center;
-    }
-
-    .code-description {
-      font-size: 0.8em;
-      text-align: center;
-      color: $gray-60;
-      padding: 10px;
-      margin-bottom: 20px;
-    }
-
-    .modal-actions {
-      margin-top: 15px;
-      display: flex;
-      flex-direction: row;
-      justify-content: flex-end;
-
-      .button {
-        padding: 5px;
-        height: 40px;
-        margin-left: 25px;
-      }
-
-      .text-button {
-        color: $gray-60;
-      }
-    }
-  }
-}
