--- app/src/pages/CreateGroup/components/VerificationModal/index.js
@@ -1,3 +1,0 @@
-import VerificationModal from './VerificationModal';
-
-export default VerificationModal;
