+++ app/src/pages/Group/Group.jsx
@@ -6,12 +6,12 @@ import Loading from '../../components/Loading';
 import PageHeader from '../../components/PageHeader';
 import Dropdown from '../../components/Dropdown';
 import Button from '../../components/Button';
+import DeleteGroupModal from '../../modals/DeleteGroupModal';
 import TopPlayerWidget from './components/TopPlayerWidget';
 import TotalExperienceWidget from './components/TotalExperienceWidget';
 import CompetitionWidget from './components/CompetitionWidget';
 import GroupInfo from './components/GroupInfo';
 import MembersTable from './components/MembersTable';
-import DeleteGroupModal from './components/DeleteGroupModal';
 import { getGroup, isFetchingMembers, isFetchingMonthlyTop } from '../../redux/selectors/groups';
 import { getGroupCompetitions } from '../../redux/selectors/competitions';
 import fetchDetailsAction from '../../redux/modules/groups/actions/fetchDetails';
